Move _nbytes to _zstdfile.py
